renamed files. Closes #53
