+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:we_find/screens/HomeScreen.dart';
+import 'package:we_find/screens/home_screen.dart';
 import 'package:we_find/styles/theme.dart' as Theme;
 
 void main() => runApp(WeFindApp());
