+++ lib/providers/lang_provider.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:we_find/model/I18NString.dart';
+import 'package:we_find/model/i18n_string.dart';
 
 /// Provider that keeps track of the language set by the user
 class LangProvider with ChangeNotifier {
