+++ lib/screens/course_detail_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:we_find/model/modelWrapped.dart';
+import 'package:we_find/model/model_wrapped.dart';
 import 'package:we_find/widgets/course_detail_widgets.dart';
 
 class CourseDetailScreen extends StatelessWidget {
