+++ test/model/i18n_string_test.dart
@@ -1,5 +1,5 @@
 import 'package:test/test.dart';
-import 'package:we_find/model/I18NString.dart';
+import 'package:we_find/model/i18n_string.dart';
 
 import 'model_matchers.dart';
