+++ test/model/model_matchers.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:we_find/model/I18NString.dart';
+import 'package:we_find/model/i18n_string.dart';
 
 class NumberOfLanguages extends CustomMatcher {
   NumberOfLanguages(matcher)
